Attempting fix death animation,  static drop item script implemented.
Static class created to drop item. Passing reference to gameObject holding item to dtop and the item to drop's location in the parents hierarchy.  Object to drop once disassociated from parent is given a Rigidbody component and a box collider. TBD about how the players can interact with the gun in this state.  Bug to be fixed: The animation controller is setup  where the appropriate animation clip is being played. The animations clip being played out looks fine in preview but very different at runtime.